Stub IndentBlockCommand.
