+++ src/indentblockcommand.js
@@ -1,0 +1,62 @@
+/**
+ * @license Copyright (c) 2003-2019, CKSource - Frederico Knabben. All rights reserved.
+ * For licensing, see LICENSE.md or https://ckeditor.com/legal/ckeditor-oss-license
+ */
+
+/**
+ * @module indent-block/indentblockcommand
+ */
+
+import Command from '@ckeditor/ckeditor5-core/src/command';
+import first from '@ckeditor/ckeditor5-utils/src/first';
+
+/**
+ * The block indentation feature.
+ *
+ * @extends module:core/plugin~Plugin
+ */
+export default class IndentBlockCommand extends Command {
+	/**
+	 * @inheritDoc
+	 */
+	refresh() {
+		this.isEnabled = this._checkEnabled();
+	}
+
+	/**
+	 * Indents or outdents (depends on the {@link #constructor}'s `indentDirection` parameter) selected list items.
+	 *
+	 * @fires execute
+	 */
+	execute() {
+		const model = this.editor.model;
+		const doc = model.document;
+
+		const itemsToChange = Array.from( doc.selection.getSelectedBlocks() );
+
+		model.change( () => {
+			for ( const item of itemsToChange ) {
+				// eslint-disable-next-line no-undef
+				console.log( 'indent block', item );
+			}
+		} );
+	}
+
+	/**
+	 * Checks whether the command can be enabled in the current context.
+	 *
+	 * @private
+	 * @returns {Boolean} Whether the command should be enabled.
+	 */
+	_checkEnabled() {
+		// Check whether any of position's ancestor is a list item.
+		const block = first( this.editor.model.document.selection.getSelectedBlocks() );
+
+		// If selection is not in a list item, the command is disabled.
+		if ( !block || !this.editor.model.schema.checkAttribute( block, 'indent' ) ) {
+			return false;
+		}
+
+		return true;
+	}
+}
